[C++ Comm] Break apart monolithic Comm unit tests
Break the unit tests into multiple files.
