+++ cpp/test/comm/unit_test_comm.bond
@@ -1,6 +1,6 @@
 
 import "bond/core/bond.bond"
-namespace unittest.rpc
+namespace unittest.comm
 
 enum Operation
 {
@@ -11,20 +11,8 @@ enum Operation
     ReturnUserError,
     ForgetCallback,
     DoubleDispatch,
-    ForwardCallback,
     FetchCallback,
     PostCallback,
-
-    PassCallbackAndReleaseBeforeResponse,
-    PassCallbackAndReleaseAfterResponse,
-    PassCallbackAndThrow,
-
-    PassCallbackAndInvokeNotifyBeforeAfterReleaseAndSchedule,
-    PassCallbackInvokeNotifyAndThrow,
-
-    PassCallbackAndInvokeUpdateBeforeAfterReleaseAndSchedule,
-    PassCallbackInvokeUpdateAndThrow,
-    PassCallbackInvoke,
 };
 
 struct Params
@@ -48,39 +36,9 @@ struct Data
     1: uint64 m_x;
 };
 
-service Listener
-{
-    nothing Notify();
-    nothing Update(Params);
-    void Invoke();
-    Data GetInstance();
-};
-
 service First
 {
     Result Method1(Params);
-
-    void MethodVoidParams(Params);
-
-    Result MethodResultVoid();
-
-    void MethodVoid();
-
-    void MethodVoidExceptionParams(Params);
-
-    Result MethodResultVoidException();
-
-    void MethodVoidException();
-
-    bond.Box<Result> MethodGenParams(Params);
-
-    Result MethodResultGen(bond.Box<Params>);
-
-    bond.Box<Result> MethodGen(bond.Box<Params>);
-
-    Result GetInstance();
-
-    nothing Event(Params);
 };
 
 service Second
@@ -88,25 +46,8 @@ service Second
     Result Method2(Params);
 };
 
-service Test
-{
-    Result TestMethod1(Params);
-    void TestMethod2();
-};
-
-service Generic<T>
-{
-    T MethodGenParams(Params);
-
-    Result MethodResultGen(T);
-        
-    T MethodGen(T);
-
-    bond.Box<T> MethodBox(bond.Box<T>);
-};
-
 struct Dummy
 {
     0: int32 int_value;
     1: string string_value;
-}
+};
