+++ cpp/test/comm/unit_test_comm_generic.bond
@@ -1,0 +1,23 @@
+
+import "unit_test_comm.bond"
+namespace unittest.comm
+
+service GenericType
+{
+    bond.Box<Result> MethodGenParams(Params);
+
+    Result MethodResultGen(bond.Box<Params>);
+
+    bond.Box<Result> MethodGen(bond.Box<Params>);
+}
+
+service Generic<T>
+{
+    T MethodGenParams(Params);
+
+    Result MethodResultGen(T);
+        
+    T MethodGen(T);
+
+    bond.Box<T> MethodBox(bond.Box<T>);
+};
