+++ cpp/test/comm/unit_test_comm_session.bond
@@ -1,0 +1,8 @@
+
+import "unit_test_comm.bond"
+namespace unittest.comm
+
+service SessionService
+{
+    Result GetInstance();
+}
